odb_backend: Implement support for the  manual reordering of odb backends
libgit2 has historically assumed that most objects for long-running
operations will be packed, therefore GIT_LOOSE_PRIORITY is set to
1 by default, and GIT_PACKED_PRIORITY is set at 2. When a client
allows libgit2 to set the default backends, it would be helpful to
be able to rearrange the default backends without having to completely
set up a new odb from scratch. This commit adds a method to set
the priority for a backend in an odb and relies on the sorting in
git_odb_add_backend to rearrange the backends.